pyproject.toml
[tool.poetry]
name = "django-recaptcha-enterprise"
version = "0.3.0"
description = ""
authors = ["rapidbounce <customer.support@rapidbounce.co>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
google-cloud-recaptcha-enterprise = "1.8.0"
grpcio = "^1.51.1"
protobuf = "3.20.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"